+++ app/src/main/java/org/oxycblt/auxio/settings/Settings.kt
@@ -19,49 +19,75 @@ package org.oxycblt.auxio.settings
 
 import android.content.Context
 import android.content.SharedPreferences
+import androidx.annotation.StringRes
 import androidx.preference.PreferenceManager
+import org.oxycblt.auxio.util.logW
+import org.oxycblt.auxio.util.unlikelyToBeNull
 
 /**
  * Abstract user configuration information. This interface has no functionality whatsoever. Concrete
  * implementations should be preferred instead.
  * @author Alexander Capehart (OxygenCobalt)
  */
-interface Settings {
-    /** Migrate any settings fields from older versions into their new counterparts. */
+interface Settings<L> {
+    /**
+     * Migrate any settings fields from older versions into their new counterparts.
+     * @throws NotImplementedError If there is nothing to migrate.
+     */
     fun migrate() {
         throw NotImplementedError()
     }
 
     /**
-     * Add a [SharedPreferences.OnSharedPreferenceChangeListener] to monitor for settings updates.
-     * @param listener The [SharedPreferences.OnSharedPreferenceChangeListener] to add.
+     * Add a listener to monitor for settings updates. Will do nothing if
+     * @param listener The listener to add.
      */
-    fun addListener(listener: SharedPreferences.OnSharedPreferenceChangeListener) {
-        throw NotImplementedError()
-    }
+    fun registerListener(listener: L)
 
     /**
-     * Unregister a [SharedPreferences.OnSharedPreferenceChangeListener], preventing any further
-     * settings updates from being sent to ti.t
+     * Unregister a listener, preventing any further settings updates from being sent to it.
+     * @param listener The listener to unregister, must be the same as the current listener.
      */
-    fun removeListener(listener: SharedPreferences.OnSharedPreferenceChangeListener) {
-        throw NotImplementedError()
-    }
+    fun unregisterListener(listener: L)
 
     /**
      * A framework-backed [Settings] implementation.
      * @param context [Context] required.
      */
-    abstract class Real(protected val context: Context) : Settings {
+    abstract class Real<L>(private val context: Context) :
+        Settings<L>, SharedPreferences.OnSharedPreferenceChangeListener {
         protected val sharedPreferences: SharedPreferences =
             PreferenceManager.getDefaultSharedPreferences(context.applicationContext)
 
-        override fun addListener(listener: SharedPreferences.OnSharedPreferenceChangeListener) {
-            sharedPreferences.registerOnSharedPreferenceChangeListener(listener)
+        /** @see [Context.getString] */
+        protected fun getString(@StringRes stringRes: Int) = context.getString(stringRes)
+
+        private var listener: L? = null
+
+        override fun registerListener(listener: L) {
+            if (this.listener == null) {
+                // Registering a listener when it was null prior, attach the callback.
+                sharedPreferences.registerOnSharedPreferenceChangeListener(this)
+            }
+            this.listener = listener
         }
 
-        override fun removeListener(listener: SharedPreferences.OnSharedPreferenceChangeListener) {
-            sharedPreferences.unregisterOnSharedPreferenceChangeListener(listener)
+        override fun unregisterListener(listener: L) {
+            if (this.listener !== listener) {
+                logW("Given listener was not the current listener.")
+            }
+            this.listener = null
+            // No longer have a listener, detach from the preferences instance.
+            sharedPreferences.unregisterOnSharedPreferenceChangeListener(this)
         }
+
+        final override fun onSharedPreferenceChanged(
+            sharedPreferences: SharedPreferences,
+            key: String
+        ) {
+            onSettingChanged(key, unlikelyToBeNull(listener))
+        }
+
+        open fun onSettingChanged(key: String, listener: L) {}
     }
 }